// src/app/api/model/applied-opportunity-statistics-dto.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AppliedOpportunityStatisticsDto { 
    done?: number;
    inProgress?: number;
    newOpportunities?: number;
    total?: number;
}